+++ dreamplace/ops/pin_pos/src/pin_pos_cuda_segment_kernel.cu
@@ -51,7 +51,7 @@ int computePinPosCudaSegmentLauncher(
 	const T* x, const T* y,
 	const T* pin_offset_x,
 	const T* pin_offset_y,
-	const long* pin2node_map,
+	const int* pin2node_map,
 	const int* flat_node2pin_map,
 	const int* flat_node2pin_start_map,
 	int num_pins,
@@ -71,7 +71,7 @@ int computePinPosGradCudaSegmentLauncher(
 	const T* x, const T* y,
 	const T* pin_offset_x,
 	const T* pin_offset_y,
-	const long* pin2node_map,
+	const int* pin2node_map,
 	const int* flat_node2pin_map,
 	const int* flat_node2pin_start_map,
 	int num_nodes,
@@ -113,7 +113,7 @@ int computePinPosGradCudaSegmentLauncher(
     	    const T* x, const T* y, \
     	    const T* pin_offset_x, \
 	        const T* pin_offset_y, \
-	        const long* pin2node_map, \
+	        const int* pin2node_map, \
 	        const int* flat_node2pin_map, \
 	        const int* flat_node2pin_start_map, \
 	        int num_pins, \
@@ -125,7 +125,7 @@ int computePinPosGradCudaSegmentLauncher(
 	        const T* x, const T* y, \
 	        const T* pin_offset_x, \
 	        const T* pin_offset_y, \
-	        const long* pin2node_map, \
+	        const int* pin2node_map, \
 	        const int* flat_node2pin_map, \
 	        const int* flat_node2pin_start_map, \
 	        int num_nodes, \
